+++ src/data-values/export/BaseConfig.js
@@ -1,8 +1,9 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { BaseStruct } from '@this/data-values';
 import { AskIf, MustBe } from '@this/typey';
+
+import { BaseStruct } from '#x/BaseStruct';
 
 
 /**