+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Breaking changes:
   * Renamed `Struct` -> `Sexp`, because it's really this project's version of
     the "sexp" concept. And for similar reasons, renamed the `type` field of it
     to `functor`. This rename also makes room for the new `Struct`-y thing.
+  * Moved `BaseConfig` here from `compy`.
   * New class `BaseStruct`, extracted from `BaseConfig`, because _most_ of what
     `BaseConfig` did was not particularly specific to configuration, per se.